fix: rename package to follow standard convention
