+++ server/shared-library/src/main/java/shopsphere/shared/SharedLibraryApplication.java
