+++ server/shared-library/src/main/java/shopsphere/shared/dto/GetUser.java
