+++ server/shared-library/src/main/java/shopsphere/shared/dto/ValidationException.java
