+++ server/shared-library/src/main/java/shopsphere/shared/enums/Role.java
@@ -1,4 +1,4 @@
-package shopsphere_shared;
+package shopsphere_shared.enums;
 
 public enum Role {
     USER,
