+++ server/shared-library/src/main/java/shopsphere/shared/exceptions/BadRequestException.java
