+++ server/shared-library/src/main/java/shopsphere/shared/exceptions/ConflictException.java
