+++ server/shared-library/src/main/java/shopsphere/shared/exceptions/ForbiddenException.java
