+++ server/shared-library/src/main/java/shopsphere/shared/handler/CustomExceptionHandler.java
