+++ server/shared-library/src/main/java/shopsphere/shared/utils/HeaderUtil.java
