+++ server/shared-library/src/test/java/shopsphere/shared/SharedLibraryApplicationTests.java
